# perf: improve rlp.decode