--- a/packages/rlp/package.json
+++ b/packages/rlp/package.json
@@ -42,7 +42,9 @@
     "src"
   ],
   "scripts": {
+    "benchmarks": "node ./benchmarks/index.js",
     "build": "../../config/cli/ts-build.sh node",
+    "build:benchmarks": "npm run build && tsc -p tsconfig.benchmarks.json",
     "clean": "../../config/cli/clean-package.sh",
     "coverage": "c8 --all --reporter=lcov --reporter=text npm run test:node",
     "examples": "tsx ../../scripts/examples-runner.ts -- rlp",
@@ -57,5 +59,8 @@
   },
   "engines": {
     "node": ">=18"
+  },
+  "devDependencies": {
+    "benchmark": "^2.1.4"
   }
 }