--- a/packages/rlp/tsconfig.benchmarks.json
+++ b/packages/rlp/tsconfig.benchmarks.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../config/tsconfig.json",
+  "compilerOptions": {
+    "lib": ["dom"],
+    "sourceMap": false,
+    "declaration": true,
+    "esModuleInterop": true,
+    "module": "esnext"
+  },
+  "include": ["benchmarks/**.ts"]
+}