Improve performance in the inner _decode function by avoiding copying of the remainder.
In the case that the final remainder is returned to the user (ie: stream=true), it is copied before returning.

Results in ~10-30x speedup

benchmarks:
before:
```
decode - list of small numbers x 3,985 ops/sec ±4.57% (42 runs sampled)
decode - list of large numbers x 3,522 ops/sec ±9.29% (32 runs sampled)
decode - list of small buffers x 2,558 ops/sec ±2.94% (41 runs sampled)
decode - list of large buffers x 0.34 ops/sec ±42.70% (5 runs sampled)
```

after:
```
decode - list of small numbers x 50,158 ops/sec ±0.33% (96 runs sampled)
decode - list of large numbers x 50,496 ops/sec ±0.44% (93 runs sampled)
decode - list of small buffers x 46,721 ops/sec ±0.29% (92 runs sampled)
decode - list of large buffers x 11.65 ops/sec ±2.61% (33 runs sampled)
```